Refactor: Modularize app.py using Flask Blueprints
I've decomposed the monolithic casino_be/app.py into multiple Blueprints for improved maintainability and scalability.

Key changes:
- Created a `routes` directory (`casino_be/routes/`) to house Blueprint files.
- Migrated functionalities into specific Blueprints:
    - `auth.py`: Authentication routes (login, register, me, etc.)
    - `user.py`: User account management (deposit, withdraw, settings)
    - `admin.py`: Admin panel routes and `is_admin` helper
    - `slots.py`: Slot machine listing, join, and spin logic
    - `blackjack.py`: Blackjack table listing, join, and game actions
    - `poker.py`: Poker table listing, join, leave, state, and game actions
    - `plinko.py`: Plinko game play route
    - `roulette.py`: Roulette game bet and history routes
    - `spacecrash.py`: Spacecrash game routes, including admin phase control
    - `baccarat.py`: Baccarat game table listing, join, and play routes (refactored from app.py)
    - `meta_game.py`: Generic game session management (e.g., end_session)
- Updated `app.py`:
    - Removed all migrated route logic.
    - `app.py` now primarily handles app initialization, configuration (Flask, DB, JWT, Limiter), registration of all Blueprints, global error handlers, and CLI commands.
    - Corrected `Flask-Limiter` initialization using `limiter.init_app(app)` to prevent circular dependencies when blueprints import the `limiter` instance.
- Moved `SATOSHIS_PER_UNIT` constant to `utils.plinko_helper.py`.
- The generic `/api/join` route was removed, deferring join logic to specific game blueprint join routes.

This refactoring enhances code organization, making it easier to manage and extend different parts of your application independently.
