+++ casino_be/routes/auth.py
@@ -1,0 +1,136 @@
+from flask import Blueprint, request, jsonify, current_app
+from flask_jwt_extended import (
+    create_access_token, create_refresh_token, jwt_required,
+    get_jwt_identity, get_jti, current_user
+)
+from datetime import datetime, timedelta, timezone
+from marshmallow import ValidationError
+
+from ..models import db, User, TokenBlacklist
+from ..schemas import UserSchema, RegisterSchema, LoginSchema
+from ..utils.bitcoin import generate_bitcoin_wallet
+from ..app import limiter # Assuming limiter can be imported directly
+
+auth_bp = Blueprint('auth', __name__, url_prefix='/api')
+
+@auth_bp.route('/me', methods=['GET'])
+@jwt_required()
+def get_current_user():
+    try:
+        user_data = UserSchema().dump(current_user)
+        current_app.logger.info(f"User profile fetched: {current_user.username} (ID: {current_user.id})")
+        return jsonify({'status': True, 'user': user_data}), 200
+    except Exception as e:
+        current_app.logger.error(f"Error fetching user profile: {str(e)}", exc_info=True)
+        return jsonify({'status': False, 'status_message': 'Failed to fetch user profile.'}), 500
+
+@auth_bp.route('/register', methods=['POST'])
+@limiter.limit("10 per hour")
+def register():
+    data = request.get_json()
+    errors = RegisterSchema().validate(data)
+    if errors:
+        return jsonify({'status': False, 'status_message': errors}), 400
+    if User.query.filter_by(username=data['username']).first():
+        return jsonify({'status': False, 'status_message': 'Username already exists'}), 409
+    if User.query.filter_by(email=data['email']).first():
+        return jsonify({'status': False, 'status_message': 'Email already exists'}), 409
+    try:
+        wallet_address = generate_bitcoin_wallet()
+        if not wallet_address:
+            return jsonify({'status': False, 'status_message': 'Failed to generate wallet address for user.'}), 500
+        new_user = User(
+            username=data['username'],
+            email=data['email'],
+            password=User.hash_password(data['password']),
+            deposit_wallet_address=wallet_address
+        )
+        db.session.add(new_user)
+        db.session.commit()
+        access_token = create_access_token(identity=new_user)
+        refresh_token = create_refresh_token(identity=new_user)
+        user_data = UserSchema().dump(new_user)
+        current_app.logger.info(f"User registered: {new_user.username} (ID: {new_user.id})")
+        return jsonify({
+            'status': True, 'user': user_data,
+            'access_token': access_token, 'refresh_token': refresh_token
+        }), 201
+    except Exception as e:
+        db.session.rollback()
+        current_app.logger.error(f"Registration failed: {str(e)}", exc_info=True)
+        return jsonify({'status': False, 'status_message': 'Registration failed.'}), 500
+
+@auth_bp.route('/login', methods=['POST'])
+@limiter.limit("10 per minute")
+def login():
+    data = request.get_json()
+    schema = LoginSchema()
+    try:
+        validated_data = schema.load(data)
+    except ValidationError as err:
+        return jsonify({'status': False, 'status_message': err.messages}), 400
+
+    user = User.query.filter_by(username=validated_data['username']).first()
+    if not user or not User.verify_password(user.password, validated_data['password']):
+        return jsonify({'status': False, 'status_message': 'Invalid credentials.'}), 401
+
+    # Update last login time
+    user.last_login_at = datetime.now(timezone.utc)
+    db.session.commit()
+
+    # Generate tokens
+    access_token = create_access_token(identity=user)
+    refresh_token = create_refresh_token(identity=user)
+    user_data = UserSchema().dump(user)
+    current_app.logger.info(f"User logged in: {user.username} (ID: {user.id})")
+    return jsonify({
+        'status': True, 'user': user_data,
+        'access_token': access_token, 'refresh_token': refresh_token
+    }), 200
+
+@auth_bp.route('/refresh', methods=['POST'])
+@jwt_required(refresh=True)
+def refresh():
+    new_access_token = create_access_token(identity=current_user)
+    current_app.logger.info(f"Token refreshed for user: {current_user.username} (ID: {current_user.id})")
+    return jsonify({'status': True, 'access_token': new_access_token}), 200
+
+@auth_bp.route('/logout', methods=['POST'])
+@jwt_required()
+def logout():
+    try:
+        auth_header = request.headers.get('Authorization')
+        if not auth_header or ' ' not in auth_header:
+             return jsonify({"status": False, "status_message": "Invalid Authorization header"}), 400
+        token = auth_header.split()[1]
+        jti = get_jti(token)
+        now = datetime.now(timezone.utc)
+        expires = timedelta(hours=1)
+        db.session.add(TokenBlacklist(jti=jti, created_at=now, expires_at=now + expires))
+        db.session.commit()
+        current_app.logger.info(f"User logged out: {current_user.username} (ID: {current_user.id})")
+        return jsonify({"status": True, "status_message": "Successfully logged out"}), 200
+    except Exception as e:
+        db.session.rollback()
+        current_app.logger.error(f"Logout failed: {str(e)}", exc_info=True)
+        return jsonify({'status': False, 'status_message': 'Logout failed.'}), 500
+
+@auth_bp.route('/logout2', methods=['POST'])
+@jwt_required(refresh=True)
+def logout2():
+    try:
+        auth_header = request.headers.get('Authorization')
+        if not auth_header or ' ' not in auth_header:
+             return jsonify({"status": False, "status_message": "Invalid Authorization header"}), 400
+        token = auth_header.split()[1]
+        jti = get_jti(token)
+        now = datetime.now(timezone.utc)
+        expires = timedelta(days=7)
+        db.session.add(TokenBlacklist(jti=jti, created_at=now, expires_at=now + expires))
+        db.session.commit()
+        current_app.logger.info(f"User refresh token invalidated: {current_user.username} (ID: {current_user.id})")
+        return jsonify({"status": True, "status_message": "Refresh token invalidated"}), 200
+    except Exception as e:
+        db.session.rollback()
+        current_app.logger.error(f"Refresh token invalidation failed: {str(e)}", exc_info=True)
+        return jsonify({'status': False, 'status_message': 'Refresh token invalidation failed.'}), 500
