+++ casino_be/routes/meta_game.py
@@ -1,0 +1,29 @@
+from flask import Blueprint, request, jsonify, current_app
+from flask_jwt_extended import jwt_required, current_user
+from datetime import datetime, timezone
+
+from ..models import db, GameSession
+from ..app import limiter # Assuming limiter can be imported directly
+
+meta_game_bp = Blueprint('meta_game', __name__, url_prefix='/api')
+
+@meta_game_bp.route('/end_session', methods=['POST'])
+@jwt_required()
+def end_session():
+    user_id = current_user.id
+    now = datetime.now(timezone.utc)
+    try:
+        active_sessions = GameSession.query.filter_by(user_id=user_id, session_end=None).all()
+        if not active_sessions:
+            return jsonify({'status': True, 'status_message': 'No active session to end'}), 200
+
+        for session in active_sessions:
+            session.session_end = now
+
+        db.session.commit()
+        current_app.logger.info(f"Ended active sessions for user {user_id}")
+        return jsonify({'status': True, 'status_message': 'Session ended successfully'}), 200
+    except Exception as e:
+        db.session.rollback()
+        current_app.logger.error(f"Failed to end session: {str(e)}", exc_info=True)
+        return jsonify({'status': False, 'status_message': 'Failed to end session.'}), 500
